feat: move code highlighter to a separate component and use it throughout the project
