+++ web-client/src/components/sources/code-view.tsx
@@ -1,14 +1,8 @@
-import { Highlighter } from "shiki";
 import { Suspense, createResource } from "solid-js";
 import { Loader } from "~/components/loader";
 import { useConnection } from "~/context/connection-provider";
-import {
-  HighlighterLang,
-  createHighlighter,
-  getHighlightedCode,
-  getText,
-} from "~/lib/code-highlight";
-import sanitizeHtml from "sanitize-html";
+import { HighlighterLang, getText } from "~/lib/code-highlight";
+import { CodeHighlighter } from "../code-higlighter";
 
 type CodeViewProps = {
   path: string;
@@ -25,38 +19,13 @@ export default function CodeView(props: CodeViewProps) {
     async (textProps) => getText(...textProps)
   );
 
-  // The used highlighter does not change at all atm so it does not need to be coupled
-  const [highlighter] = createResource(() => createHighlighter());
-
-  const html = (
-    text: string | undefined,
-    highlighter: Highlighter | undefined,
-    lang: HighlighterLang,
-    highlightedLine?: number
-  ) => {
-    if (!text || !highlighter) return undefined;
-    const raw = getHighlightedCode([text, highlighter, lang, highlightedLine]);
-
-    return sanitizeHtml(raw, {
-      allowedTags: ["pre", "code", "span"],
-      allowedAttributes: {
-        pre: ["class"],
-        span: ["class", "style"],
-      },
-    });
-  };
-
   return (
     <div class="min-h-full h-max min-w-full w-max bg-black bg-opacity-50">
       <Suspense fallback={<Loader />}>
-        <div
-          //eslint-disable-next-line solid/no-innerhtml
-          innerHTML={html(
-            text(),
-            highlighter(),
-            props.lang,
-            props.highlightedLine
-          )}
+        <CodeHighlighter
+          text={text()}
+          lang={props.lang}
+          highlightedLine={props.highlightedLine}
         />
       </Suspense>
     </div>
